// chessklub1-main/src/app/types/events.ts
export interface TournamentCard {
  id: string;
  category: string;
  date: string;
  title: string;
  description: string;
  price: string;
  icon: string;
  gradientFrom: string;
  gradientTo: string;
}

export interface EventCard {
  id: string;
  category: string;
  title: string;
  description: string;
  price: string;
  gradientFrom: string;
  gradientTo: string;
}